--- specification/watcher/execute_watch/WatchRecord.ts
@@ -1,40 +1,0 @@
-/*
- * Licensed to Elasticsearch B.V. under one or more contributor
- * license agreements. See the NOTICE file distributed with
- * this work for additional information regarding copyright
- * ownership. Elasticsearch B.V. licenses this file to you under
- * the Apache License, Version 2.0 (the "License"); you may
- * not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-import { Dictionary } from '@spec_utils/Dictionary'
-import { UserDefinedValue } from '@spec_utils/UserDefinedValue'
-import { ConditionContainer } from '@watcher/_types/Conditions'
-import { InputContainer } from '@watcher/_types/Input'
-import { Id, Metadata } from '@_types/common'
-import { ActionExecutionState } from './ActionExecutionState'
-import { ExecutionResult } from './ExecutionResult'
-import { TriggerEventResult } from './TriggerEventResult'
-
-export class WatchRecord {
-  condition: ConditionContainer
-  input: InputContainer
-  messages: string[]
-  metadata: Metadata
-  node: string
-  result: ExecutionResult
-  state: ActionExecutionState
-  trigger_event: TriggerEventResult
-  user: string
-  watch_id: Id
-}
