+++ specification/watcher/execute_watch/WatcherExecuteWatchResponse.ts
@@ -18,7 +18,7 @@
  */
 
 import { Id } from '@_types/common'
-import { WatchRecord } from './WatchRecord'
+import { WatchRecord } from './types'
 
 export class Response {
   body: { _id: Id; watch_record: WatchRecord }
