+++ specification/watcher/stats/types.ts
@@ -17,6 +17,10 @@
  * under the License.
  */
 
+import { Id } from '@_types/common'
+import { long } from '@_types/Numeric'
+import { DateString } from '@_types/Time'
+
 export enum ExecutionPhase {
   awaits_execution = 0,
   started = 1,
@@ -26,4 +30,44 @@ export enum ExecutionPhase {
   watch_transform = 5,
   aborted = 6,
   finished = 7
+}
+
+export enum WatcherState {
+  stopped = 0,
+  starting = 1,
+  started = 2,
+  stopping = 3
+}
+
+export class WatcherNodeStats {
+  current_watches?: WatchRecordStats[]
+  execution_thread_pool: ExecutionThreadPool
+  queued_watches?: WatchRecordQueuedStats[]
+  watch_count: long
+  watcher_state: WatcherState
+  node_id: Id
+}
+
+export class ExecutionThreadPool {
+  max_size: long
+  queue_size: long
+}
+
+export enum WatcherMetric {
+  '_all' = 0,
+  'queued_watches' = 1,
+  'current_watches' = 2,
+  'pending_watches' = 3
+}
+
+export class WatchRecordQueuedStats {
+  execution_time: DateString
+}
+
+export class WatchRecordStats extends WatchRecordQueuedStats {
+  execution_phase: ExecutionPhase
+  triggered_time: DateString
+  executed_actions?: Array<string>
+  watch_id: Id
+  watch_record_id: Id
 }
